Improved callbacks
Using a more intuitive API to set on-libzim-written callback to Item.
Both `add_item_for()` and `add_item()` now gets a `callback` argument that can be
either a single callable or a tuple of callable + args.

This is now implemented as a weakref finalize on the Item which will be triggered
by python's garbage collector once the libzim gets rid of the content provider: we
keep a ref from CP to Item so that we can attach our callback to the Item.
Doing this fixes an important memory leak by not using the __del__

`add_item_for()`'s `remove_fpath` behavior has been changed to use this new callback
mechanism and account for the case where both this and a callback are expected