Fix handling of tag list in Creator.add_metadata (fix #125)
This fixes issue #125. Previously, the type annotations of Creator.config_metadata()
allowed "Tags" to be a list of strings, but didn't handle the conversion to a string,
which resulted in python-libzim raising an error.

This commit modifies Creator.add_metadata() to accept a list of strings and, if the
key of the metadata is "Tags", join them into a single string. A regression test is
included.