Mutualize test code base + add pyright hint
pyright ignore is mandatory because the type checker has no idea about
the impact of validate_metadata/validate_tag. This might be fixed by a
shared logic and a TypeGuard but is not available until Python 3.10 ;
other solution would be to transfer metadata to a way more typed
container after validation.