Add back a runinstalled marker for ogvjs tests
Before 3.3, there was a --runinstalled switch for tests to disable by
default the execution of some tests which were deemed to work only if
the package tools were installed.

In 3.3.0, we assumed this switch was not necessary anymore before with
new Python bootstrap convention, we are always in installed mode.

This statement was indeed wrong because some persons are building and
testing without using our build chain (e.g. they build and test directly
from sdist).

This commit adds back the --runinstalled switch for tests and
corresponding pytest marker to mark tests that are going to work only if
package tools are installed.